+++ CHANGELOG.md
@@ -18,6 +18,7 @@ make timgen-build
    - `native_crs` with code of native CRS in form "EPSG:&lt;code&gt;", e.g. "EPSG:4326"
    - `native_bounding_box` with coordinates in native CRS  [minx, miny, maxx, maxy]
 - [#64](https://github.com/LayerManager/layman/issues/64) Layman supports import of vector layers in "EPSG:3034", "EPSG:3035", "EPSG:5514", "EPSG:32633", "EPSG:32634" and "EPSG:3059".
+- [#519](https://github.com/LayerManager/layman/issues/64) Endpoints [GET Layers](doc/rest.md#get-layers), [GET Workspace Layers](doc/rest.md#get-workspace-layers), [GET Maps](doc/rest.md#get-maps), [GET Workspace Maps](doc/rest.md#get-workspace-maps) support new query parameter *bbox_filter_crs*. 
 - [#64](https://github.com/LayerManager/layman/issues/64) Layer thumbnails are generated in native CRS of the layer.
 - [#64](https://github.com/LayerManager/layman/issues/64) WMS proxy was added to [WMS endpoint](doc/endpoints.md#web-map-service). In case of some special WMS GetMap requests, it adds `buffer` parameter to the request to fix some GeoServer issues or change request CRS.
 - [#64](https://github.com/LayerManager/layman/issues/64) For layers in `EPSG:5514` and WFS requests in `CRS:84`, the features may have wrong coordinates by hundreds of meters. For requests in `EPSG:4326`, coordinates are correct.